requirements.txt
autopep8==1.4.4
pep8==1.7.1
pycodestyle==2.5.0
pytz==2018.5
sqlparse==0.2.4
django-naomi
django
djangorestframework
django-rest-auth
django-allauth
django-model-utils
django-js-reverse
django-import-export
python-decouple
psycopg2
brotlipy
django-log-request-id
dj-database-url
gunicorn
whitenoise
psutil
ipython
sentry-sdk
requests
django-cors-headers==3.4.0